Break down ClientChatPage into smaller components
- Created `ChatPageLoading`, `ChatMessages`, `ChatModuleHeader`, `WelcomeMessage`, `CodeBlock` to Chat directory for better structure.
- Reworked loading screen for module page to use loading animation
