+++ app/[moduleName]/chat/page.tsx
@@ -4,8 +4,8 @@ import { useEffect, useState } from "react";
 import { notFound, useParams, useRouter } from "next/navigation";
 import { useUser } from "@clerk/nextjs";
 import { generateId, decodeModuleSlug } from "@/lib/utils";
-import ClientChatPage from "@/components/Main/ClientChatPage";
-import { ChatPageLoading } from "@/components/Main/ClientChatPage";
+import ClientChatPage from "@/components/Chat/ClientChatPage";
+import { ChatPageLoading } from "@/components/Chat/ChatPageLoading";
 import { useMemo } from "react";
 import { Message } from "@ai-sdk/react";
 import { createModuleWelcomeMessage } from "@/lib/prompts";
