+++ components/Chat/ClientChatPage.tsx
@@ -1,0 +1,361 @@
+"use client";
+
+import React from "react";
+import { Loader2 } from "lucide-react";
+import { Message, useChat } from "@ai-sdk/react";
+import { toast } from "sonner";
+import { ModuleWithResources } from "@/lib/actions";
+import { useRouter } from "next/navigation";
+import { encodeModuleSlug } from "@/lib/utils";
+import Header from "../Main/Header";
+
+// Import the extracted components
+import ChatMessages from "./ChatMessages";
+import ChatInput from "./ChatInput";
+import ChatModuleHeader from "./ChatModuleHeader";
+import WelcomeScreen from "./WelcomeScreen";
+
+/**
+ * Renders a full-screen chat interface with message history, input, and an optional module header.
+ *
+ * This component manages the chat session by handling message submission with optimistic UI updates,
+ * dynamically updating the chat model based on API responses, and adjusting the browser URL for authenticated users.
+ * It also retrieves a session ID for anonymous users and provides clipboard functionality for copying responses.
+ *
+ * @param initialModuleDetails - Optional module details used to display context-specific header information.
+ * @param chatId - Unique identifier for the chat session.
+ * @param initialMessages - Array of messages to initialize the chat conversation.
+ * @param isAuthenticated - Indicates whether the user is authenticated (defaults to true).
+ * @param initialTitle - Optional initial title for the chat.
+ * @param forceOldest - When true, ensures this chat appears as the oldest in history (for unauthenticated users).
+ *
+ * @returns The rendered chat interface as a React element.
+ */
+export default function ClientChatPage({
+  initialModuleDetails,
+  chatId,
+  initialMessages = [],
+  isAuthenticated = true,
+  initialTitle,
+  forceOldest = false,
+}: {
+  initialModuleDetails?: ModuleWithResources | null;
+  chatId: string;
+  initialMessages?: Message[];
+  isAuthenticated?: boolean;
+  initialTitle?: string;
+  forceOldest?: boolean;
+}) {
+  const [showLogo, setShowLogo] = React.useState(false);
+
+  // After mounting, we have access to the theme
+  React.useEffect(() => {
+    // Add a small delay before showing the logo to prevent flashing
+    const timer = setTimeout(() => {
+      setShowLogo(true);
+    }, 50);
+    return () => clearTimeout(timer);
+  }, []);
+
+  // These states are used for initial values and potential future updates
+  const [activeModule] = React.useState<string | null>(
+    initialModuleDetails?.id ?? null
+  );
+  const [moduleDetails] = React.useState<ModuleWithResources | null>(
+    initialModuleDetails ?? null
+  );
+  const router = useRouter();
+
+  // Get the session ID from local storage for anonymous users
+  const [sessionId, setSessionId] = React.useState<string | null>(null);
+
+  // Get access to the sidebar's addOptimisticChat function
+  // This allows us to update the chat history immediately when sending a message
+  const sidebarChatUpdater = React.useRef<
+    | ((
+        title: string,
+        moduleId: string | null,
+        forceOldest?: boolean
+      ) => string)
+    | null
+  >(null);
+
+  // Connect to the optimistic chat updater if available in the global window object
+  React.useEffect(() => {
+    if (typeof window !== "undefined") {
+      const win = window as unknown as {
+        __sidebarChatUpdater?: (
+          title: string,
+          moduleId: string | null,
+          forceOldest?: boolean
+        ) => string;
+      };
+      if (win.__sidebarChatUpdater) {
+        sidebarChatUpdater.current = win.__sidebarChatUpdater;
+      }
+    }
+  }, []);
+
+  // Load sessionId from localStorage on component mount (client-side only)
+  React.useEffect(() => {
+    if (typeof window !== "undefined" && !isAuthenticated) {
+      // Import the session utility dynamically since it's a client-side module
+      import("@/lib/session").then(({ getOrCreateSessionIdClient }) => {
+        const storedSessionId = getOrCreateSessionIdClient();
+        if (storedSessionId) {
+          setSessionId(storedSessionId);
+        }
+      });
+    }
+  }, [isAuthenticated]);
+
+  // For the welcome chat, add it to history for anonymous users on initial load
+  React.useEffect(() => {
+    if (
+      !isAuthenticated &&
+      sessionId &&
+      initialTitle &&
+      sidebarChatUpdater.current &&
+      // Specifically check for the welcome chat
+      (forceOldest ||
+        chatId === "welcome-chat" ||
+        chatId.startsWith("welcome-chat-"))
+    ) {
+      // Add the welcome chat to the sidebar for anonymous users
+      sidebarChatUpdater.current(initialTitle, activeModule, true);
+
+      // If this is the welcome chat for first-time users, make sure they see the welcome message
+      if (
+        (chatId === "welcome-chat" || chatId.startsWith("welcome-chat-")) &&
+        window.location.pathname === "/chat/welcome"
+      ) {
+        console.log("Welcome chat loaded for first-time user");
+      }
+    }
+  }, [
+    isAuthenticated,
+    sessionId,
+    initialTitle,
+    forceOldest,
+    activeModule,
+    chatId,
+  ]);
+
+  // Reference to the scroll container, but don't auto-scroll
+  const scrollContainerRef = React.useRef<HTMLDivElement>(null);
+
+  // Default model name - we'll try to retrieve it dynamically if possible
+  const [modelName, setModelName] = React.useState<string>("Gemini 2.0 Flash");
+
+  // Memoize the chat options to prevent unnecessary re-initialization
+  const chatOptions = React.useMemo(
+    () => ({
+      api: "/api/chat",
+      id: chatId,
+      initialMessages,
+      body: {
+        moduleId: activeModule,
+        chatId: chatId,
+        isAuthenticated: isAuthenticated,
+        sessionId: !isAuthenticated ? sessionId : undefined,
+        title: initialTitle,
+        forceOldest: forceOldest,
+      },
+      onResponse: (response: Response) => {
+        // Try to extract model information from headers if available
+        const modelHeader = response.headers.get("x-model-used");
+        if (modelHeader) {
+          setModelName(modelHeader);
+        }
+
+        // Only update URL with chat ID for authenticated users
+        if (isAuthenticated) {
+          // Update URL to include the chat ID after user sends a message
+          const currentPath = window.location.pathname;
+
+          // Make sure chatId is defined before updating URL
+          if (!chatId) {
+            console.error("Error: chatId is undefined");
+            return;
+          }
+
+          if (currentPath === "/chat" && !activeModule) {
+            // Only update URL if we're on the main chat page and not in a module
+            router.replace(`/chat/${chatId}`, { scroll: false });
+          } else if (activeModule && moduleDetails) {
+            // For module chats, update URL to the proper format
+            const encodedName = encodeModuleSlug(moduleDetails.name);
+
+            // More flexible path handling for module chats
+            // This will work for both initial module chat pages and existing chat pages
+            if (currentPath.includes(`/${encodedName}/chat`)) {
+              // Check if we need to update the URL (if it doesn't already contain the chat ID)
+              if (!currentPath.endsWith(`/${chatId}`)) {
+                const newPath = `/${encodedName}/chat/${chatId}`;
+                console.log(`Updating URL to: ${newPath}`);
+                router.replace(newPath, { scroll: false });
+              }
+            }
+          }
+        }
+      },
+      onError: (error: Error) => {
+        toast.error("Error: " + (error.message || "Failed to send message"));
+      },
+    }),
+    [
+      chatId,
+      initialMessages,
+      activeModule,
+      isAuthenticated,
+      moduleDetails,
+      sessionId,
+      router,
+      initialTitle,
+      forceOldest,
+    ]
+  );
+
+  const {
+    messages,
+    input,
+    handleInputChange,
+    handleSubmit,
+    isLoading: chatLoading,
+  } = useChat(chatOptions);
+
+  const [copiedMessageId, setCopiedMessageId] = React.useState<string | null>(
+    null
+  );
+
+  const copyToClipboard = React.useCallback(
+    (text: string, messageId: string) => {
+      navigator.clipboard.writeText(text).then(() => {
+        setCopiedMessageId(messageId);
+        setTimeout(() => setCopiedMessageId(null), 2000);
+      });
+    },
+    []
+  );
+
+  const navigateToModuleDetails = React.useCallback(() => {
+    if (moduleDetails) {
+      const encodedName = encodeModuleSlug(moduleDetails.name);
+      router.push(`/modules/${encodedName}`);
+    }
+  }, [moduleDetails, router]);
+
+  // Add a callback for when the first message is sent
+  const handleFirstMessageSent = React.useCallback(
+    (message: string) => {
+      // Update the chat history optimistically when the first message is sent
+      if (sidebarChatUpdater.current && messages.length === 0) {
+        // Create a chat title from the first message
+        const chatTitle =
+          message.substring(0, 30) + (message.length > 30 ? "..." : "");
+        sidebarChatUpdater.current(chatTitle, activeModule, forceOldest);
+
+        // Log for anonymous users
+        if (!isAuthenticated && sessionId) {
+          console.log(
+            `Creating optimistic chat for anonymous user with sessionId: ${sessionId}`
+          );
+        }
+      }
+    },
+    [messages.length, isAuthenticated, activeModule, sessionId, forceOldest]
+  );
+
+  // Handle form submission with optimized event handler
+  const handleFormSubmit = React.useCallback(
+    (e: React.FormEvent) => {
+      e.preventDefault();
+      if (input.trim() && !chatLoading) {
+        // Track if this is the first message for the optimistic UI update
+        const isFirstMessage = messages.length === 0;
+
+        // Submit the form
+        handleSubmit(e);
+
+        // If this is the first message, trigger the optimistic UI update
+        if (isFirstMessage) {
+          handleFirstMessageSent(input.trim());
+        }
+      }
+    },
+    [input, chatLoading, handleSubmit, messages.length, handleFirstMessageSent]
+  );
+
+  // Handle keyboard event
+  const handleKeyDown = React.useCallback(
+    (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
+      if (e.key === "Enter" && !e.shiftKey) {
+        e.preventDefault();
+        if (input.trim() && !chatLoading) {
+          const form = e.currentTarget.form;
+          if (form) form.requestSubmit();
+        }
+      }
+    },
+    [input, chatLoading]
+  );
+
+  return (
+    <div className="flex h-screen flex-col overflow-hidden">
+      {/* Header component */}
+      <Header />
+
+      {/* Main Content Area with Scrollbar - Make it span the full page */}
+      <div
+        ref={scrollContainerRef}
+        className={`flex-1 flex flex-col ${
+          messages.length > 0 ? "overflow-y-auto" : "overflow-hidden"
+        } pr-0 scroll-smooth scrollbar-smooth custom-scrollbar`}
+      >
+        {/* Chat Header - Now inside the scrollable area */}
+        {moduleDetails && (
+          <ChatModuleHeader
+            moduleDetails={moduleDetails}
+            navigateToModuleDetails={navigateToModuleDetails}
+          />
+        )}
+
+        {/* Chat content centered container */}
+        <div className="flex-1">
+          {/* Ensure chat thread has same width as input by using identical container classes */}
+          <div className="max-w-3xl mx-auto w-full transition-all duration-200">
+            {/* Use the same padding as the input container */}
+            <div className="px-0">
+              {messages.length === 0 ? (
+                <WelcomeScreen showLogo={showLogo} />
+              ) : (
+                <div className="flex flex-col space-y-8 pt-14 px-2 pb-8">
+                  <ChatMessages
+                    messages={messages}
+                    copiedMessageId={copiedMessageId}
+                    copyToClipboard={copyToClipboard}
+                    modelName={modelName}
+                  />
+                </div>
+              )}
+              {chatLoading && (
+                <div className="pl-8 pr-6 mt-4">
+                  <Loader2 className="h-5 w-5 animate-spin" />
+                </div>
+              )}
+            </div>
+          </div>
+        </div>
+
+        {/* Input form - Now using the ChatInput component */}
+        <ChatInput
+          input={input}
+          handleInputChange={handleInputChange}
+          handleFormSubmit={handleFormSubmit}
+          chatLoading={chatLoading}
+          handleKeyDown={handleKeyDown}
+        />
+      </div>
+    </div>
+  );
+}
